+++ docs/changelog.md
@@ -3,6 +3,7 @@
 - Deprecate `create_ambiorix`: moving to [ambiorix.generator](https://github.com/JohnCoene/ambiorix.generator) package.
 - Deprecate `add_template`: moving to [ambiorix.generator](https://github.com/JohnCoene/ambiorix.generator) package.
 - Added `all` method to define route and handler for all methods `GET`, `POST`, `PUT`, `DELETE`, and `PATCH`.
+- The `use` method now accepts a function which is run every time the server receives a request.
 
 # ambiorix 1.0.0